requirements.txt
astrapy~=1.5.2
colorama~=0.4.6
python-dotenv~=1.0.1
powerschool~=4.0.1
chromadb~=0.6.3